implement plot_density to bokeh
