+++ arviz/plots/backends/matplotlib/mpl_compareplot.py
@@ -1,3 +1,4 @@
+"""Matplotlib Compareplot."""
 import matplotlib.pyplot as plt
